Rename Core Worker project and folder name
