+++ src/dotnet/CoreWorker/.dockerignore
