+++ src/dotnet/CoreWorker/CoreWorker.csproj
@@ -4,7 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
-    <UserSecretsId>dotnet-CoreWorkerService-66b793a2-9018-4ae5-9ed4-b893e8a3e240</UserSecretsId>
+    <UserSecretsId>dotnet-CoreWorker-66b793a2-9018-4ae5-9ed4-b893e8a3e240</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
 	<GenerateDocumentationFile>True</GenerateDocumentationFile>
   </PropertyGroup>
