+++ src/dotnet/CoreWorker/Dockerfile
@@ -13,15 +13,15 @@ RUN dotnet build "dotnet/Common/Common.csproj" -c Release -o /app/build
 RUN dotnet restore "dotnet/Core/Core.csproj"
 RUN dotnet build "dotnet/Core/Core.csproj" -c Release -o /app/build
 
-RUN dotnet restore "dotnet/CoreWorkerService/CoreWorkerService.csproj"
-RUN dotnet build "dotnet/CoreWorkerService/CoreWorkerService.csproj" -c Release -o /app/build
+RUN dotnet restore "dotnet/CoreWorker/CoreWorker.csproj"
+RUN dotnet build "dotnet/CoreWorker/CoreWorker.csproj" -c Release -o /app/build
 
 FROM build AS publish
 RUN dotnet publish "dotnet/Common/Common.csproj" -c Release -o /app/publish /p:UseAppHost=false
 RUN dotnet publish "dotnet/Core/Core.csproj" -c Release -o /app/publish /p:UseAppHost=false
-RUN dotnet publish "dotnet/CoreWorkerService/CoreWorkerService.csproj" -c Release -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "dotnet/CoreWorker/CoreWorker.csproj" -c Release -o /app/publish /p:UseAppHost=false
 
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "CoreWorkerService.dll"]
+ENTRYPOINT ["dotnet", "CoreWorker.dll"]
