+++ src/dotnet/CoreWorker/appsettings.json
